+++ R/zzz_get_endpoint.R
@@ -55,6 +55,13 @@ get_endpoint <- function(endpoint = c("benthicattributes", "choices", "fishfamil
 }
 
 lookup_choices <- function(results, endpoint, endpoint_type = "main") {
+
+  endpoint <- stringr::str_remove(endpoint, "csv") %>%
+    basename()
+  if (!endpoint %in% tested_endpoints) {
+    browser()
+  }
+
   url <- base_url
 
   if (nrow(results) == 0) {
@@ -163,6 +170,7 @@ lookup_variable <- function(.data, choices, variable) {
 }
 
 construct_endpoint_columns <- function(x, endpoint) {
+  browser()
   dplyr::select(x, mermaid_endpoint_columns[[endpoint]])
 }
 
@@ -180,6 +188,8 @@ remove_blacklist_endpoint_columns <- function(res, endpoint, nested = NA_charact
 }
 
 strip_name_suffix <- function(results, endpoint, covariates = FALSE) {
+  endpoint <- stringr::str_remove(endpoint, "csv") %>%
+    basename()
   if (!endpoint %in% tested_endpoints) {
     browser()
   }
@@ -210,29 +220,26 @@ strip_name_suffix <- function(results, endpoint, covariates = FALSE) {
 
 allowed_ids <- function(endpoint, covariates = FALSE) {
   ids <- c("project_id", "sample_event_id", "sample_unit_id", "provider_id")
-
-  if (endpoint %in% names(mermaid_endpoint_columns)) {
-    # TODO -> this is explicit white listing, so will need to move away from this
-    if (any(stringr::str_ends(mermaid_endpoint_columns[[endpoint]], "_id"))) {
-      browser()
-      ids <- c(
-        ids,
-        mermaid_endpoint_columns[[endpoint]][grepl(
-          "_id$",
-          mermaid_endpoint_columns[[endpoint]]
-        )]
-      )
-    }
-  }
-
-  if (covariates) {
-    ids <- c(ids, "site_id")
-  }
+#
+#   browser()
+#
+#   if (endpoint %in% names(mermaid_endpoint_columns)) {
+#     # TODO -> this is explicit white listing, so will need to move away from this
+#     if (any(stringr::str_ends(mermaid_endpoint_columns[[endpoint]], "_id"))) {
+#       browser()
+#       ids <- c(
+#         ids,
+#         mermaid_endpoint_columns[[endpoint]][grepl(
+#           "_id$",
+#           mermaid_endpoint_columns[[endpoint]]
+#         )]
+#       )
+#     }
+#   }
+#
+#   if (covariates) {
+#     ids <- c(ids, "site_id")
+#   }
 
   unique(ids)
 }
-
-# Defined in respective function files
-mermaid_endpoint_columns <- list(
-  projects = projects_columns
-)